Add linting and test harness to the codebase
We add tox, and various linting options to the codebase to provide
a baseline for future code-review.

A placeholder unit test has been added to the repository to begin
the process of testing functions as we evolve the code-base.

NB. We will need to change the name test_placeholder.py as we begin
to do that.