Create a package handler including bagit
We add bagging to create_package. Along the way we make sure data
is only stored temporarily using mkdtemp which is Python's most
secure way to create a temporary directory that is removed after
being used.

We begin to experiment with implementing a bag profile for Arkly
that is named using a ULID and contains a hidden ULID folder
storing some metadata about the package.