Add Arkly added value endpoints to Arkly tag
These specific endpoints perform Arkly specific functions, i.e.
packaging content into a bag, or validating a bag. They're not
generic to Arweave so are separated at least for now. We also add
some documentation to the tags_metadata. We remove the docs links
as they are not currently active.