Add workflows
Unit test and linting workflows are added.

Tests are configured to run remotely without the need for files to
exist logically on disk. We have implemented mocks to do this.

The testing environment has been updated as part of this commit with
one ignore rule added for W3101 (missing-timeout) which seems to
have been added between pylint 2.14 and 2.15.