Provide GET endpoints for Arweave wallet balances
POST endpoints already existed for wallet balances but required the
Arweave keyfile. We provide GET endpoints here that take a wallet
address as an argument and return the detail to the caller.

Wallet balances are returned in Ar and Winston and helper functions
for those units have been modified to return floating point numbers
and integers for the values respectively.