Simplify imports in api.py
We were importing helper functions one by one which was getting
painful to extend.

Multiple imports have been removed to a single module import and
callers have been updated accordingly.

The change triggers W0212 (access to protected members). Using the
leading underscore was done as a convenience as naming API entry
points, entry-functions, and helper functions with FastAPI gets
complicated, and/or messy. We went with simple using an underscore
to allow duplication of names between the primary api entry point
and the helper. This can be changed in future, e.g. by adding another
suffix and/or prefix.