Remove database logging middleware
A database was being used to log API activity but in practice because
of the additional demands of configuring postgres it wasn't being
configured in production. The pattern can be observed by
investigating or reverting this commit in future releases and its
model adopted with clearer user requirements. This commit allows the
app to run more standalone and removes some complexity per request
made to the API.